+++ extensions/jupiter-aggregator/src/api.ts
@@ -58,7 +58,7 @@ export function fetchTokenList() {
   });
 }
 
-/* export const compareTokens = async (token1: string, amount: any, token2: string) => {
+export const compareTokens = async (token1: string, amount: any, token2: string) => {
   try {
     const res = await axios.get(`https://price.jup.ag/v4/price?ids=${token1}&vsToken=${token2}`);
     const result = amount * res.data.data[token1].price;
@@ -70,16 +70,4 @@ export function fetchTokenList() {
       message: "Check network connection:" + error,
     });
   }
-}; */
-
-export function compareTokens<T extends string>(token1: T, token2: string) {
-  return useFetch<PriceResponse<T>>(`https://price.jup.ag/v4/price?ids=${token1}&vsToken=${token2}`, {
-    onError: (error: Error) => {
-      showToast({
-        style: Toast.Style.Failure,
-        title: "Failed to compare tokens",
-        message: "Check network connection:" + error,
-      });
-    },
-  });
-}
+};